Split app in three files
