+++ benchmarks/foveation/CMakeLists.txt
@@ -1,4 +1,4 @@
-# Copyright 2023 Google LLC
+# Copyright 2024 Google LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -11,7 +11,6 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-cmake_minimum_required(VERSION 3.0 FATAL_ERROR)
 
 project(foveation)
 
@@ -19,6 +18,8 @@ add_vk_sample(
     NAME ${PROJECT_NAME}
     SOURCES
       "main.cpp"
+      "FoveationBenchmarkApp.h"
+      "FoveationBenchmarkApp.cpp"
     DEPENDENCIES
       "shader_foveation_benchmark"
       "shader_foveation_benchmark_shading_rate_khr"
